refactor : pr 리뷰 관련 수정
