+++ src/main/java/com/example/demo/domain/member/service/MemberQueryService.java
@@ -1,32 +1,11 @@
 package com.example.demo.domain.member.service;
 
-import com.example.demo.common.annotation.UseCase;
-import com.example.demo.common.exception.ErrorStatus;
-import com.example.demo.common.exception.GeneralException;
 import com.example.demo.domain.member.entity.Member;
-import com.example.demo.domain.member.repository.MemberRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.transaction.annotation.Transactional;
 
-@UseCase
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class MemberQueryService {
+public interface MemberQueryService {
+    Member getByKakaoEmail(String kakaoEmail);
 
-    private final MemberRepository memberRepository;
+    Member getMemberByUsername(String username);
 
-    public Member getByKakaoEmail(String kakaoEmail) {
-        return memberRepository.findByKakaoEmail(kakaoEmail)
-                .orElseThrow(() -> new GeneralException(ErrorStatus.MEMBER_NOT_FOUND));
-    }
-
-    public Member getMemberByUsername(String username) {
-        return memberRepository.findByUsername(username)
-                .orElseThrow(() -> new GeneralException(ErrorStatus.MEMBER_NOT_FOUND));
-    }
-
-    public Member getMemberById(Long memberId) {
-        return memberRepository.findById(memberId)
-                .orElseThrow(() -> new GeneralException(ErrorStatus.MEMBER_NOT_FOUND));
-    }
+    Member getMemberById(Long memberId);
 }
