+++ src/main/java/com/example/demo/security/controller/TokenApiController.java
@@ -4,6 +4,8 @@ import com.example.demo.api.common.dto.ApiResponseDto;
 import com.example.demo.common.annotation.DisableSwaggerSecurity;
 import com.example.demo.common.exception.ErrorStatus;
 import com.example.demo.common.exception.GeneralException;
+import com.example.demo.common.util.CookieUtil;
+import com.example.demo.security.filter.JwtAuthenticationFilter;
 import com.example.demo.security.jwt.dto.AccessTokenResponse;
 import com.example.demo.security.jwt.dto.JwtToken;
 import com.example.demo.security.jwt.service.TokenService;
@@ -17,6 +19,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.bind.annotation.*;
 
+import static com.example.demo.common.consts.StaticVariable.REFRESH_TOKEN_COOKIE;
+
 @Tag(name = "Token API", description = "토큰 API")
 @ApiResponse(responseCode = "2000", description = "성공")
 @RequiredArgsConstructor
@@ -24,8 +28,7 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 public class TokenApiController {
 
-    private static final String REFRESH_TOKEN_COOKIE = "refreshToken";
-
+    private final CookieUtil cookieUtil;
     private final TokenService tokenService;
 
     @Operation(
@@ -37,7 +40,7 @@ public class TokenApiController {
     public ApiResponseDto<AccessTokenResponse>  login(@RequestParam String kakaoEmail,
                                                      HttpServletResponse response) {
         JwtToken jwtToken = tokenService.login(kakaoEmail);
-        setTokenResponse(response, jwtToken);
+        cookieUtil.setTokenResponse(response, jwtToken);
         return ApiResponseDto.onSuccess(new AccessTokenResponse(jwtToken.getAccessToken()));
     }
 
@@ -49,9 +52,9 @@ public class TokenApiController {
     @PostMapping("/reissue")
     public ApiResponseDto<AccessTokenResponse> reissue(HttpServletRequest request,
                                                        HttpServletResponse response) {
-        String refreshToken = extractRefreshTokenFromCookie(request);
+        String refreshToken = cookieUtil.extractRefreshTokenFromCookie(request);
         JwtToken jwtToken = tokenService.issueTokens(refreshToken);
-        setTokenResponse(response, jwtToken);
+        cookieUtil.setTokenResponse(response, jwtToken);
         return ApiResponseDto.onSuccess(new AccessTokenResponse(jwtToken.getAccessToken()));
     }
 
@@ -61,43 +64,11 @@ public class TokenApiController {
     )
     @PostMapping("/logout")
     public ApiResponseDto<Void> logout(HttpServletRequest request, HttpServletResponse response) {
-        String refreshToken = extractRefreshTokenFromCookie(request);
+        String refreshToken = cookieUtil.extractRefreshTokenFromCookie(request);
         tokenService.logout(refreshToken);
-        expireRefreshTokenCookie(response);
+        cookieUtil.expireRefreshTokenCookie(response);
         return ApiResponseDto.onSuccess(null);
     }
 
-    /** AccessToken을 헤더에, RefreshToken을 HttpOnly Cookie에 세팅한다. */
-    private void setTokenResponse(HttpServletResponse response, JwtToken jwtToken) {
-        response.setHeader(HttpHeaders.AUTHORIZATION, "Bearer " + jwtToken.getAccessToken());
-        response.addCookie(createRefreshTokenCookie(jwtToken.getRefreshToken()));
-    }
 
-    private Cookie createRefreshTokenCookie(String refreshToken) {
-        Cookie cookie = new Cookie(REFRESH_TOKEN_COOKIE, refreshToken);
-        cookie.setHttpOnly(true);
-        cookie.setPath("/");
-        cookie.setMaxAge(7 * 24 * 60 * 60); // 7일
-        return cookie;
-    }
-
-    /** 만료된 빈 Cookie를 덮어써서 브라우저의 RefreshToken Cookie를 삭제한다. */
-    private void expireRefreshTokenCookie(HttpServletResponse response) {
-        Cookie cookie = new Cookie(REFRESH_TOKEN_COOKIE, null);
-        cookie.setHttpOnly(true);
-        cookie.setPath("/");
-        cookie.setMaxAge(0);
-        response.addCookie(cookie);
-    }
-
-    private String extractRefreshTokenFromCookie(HttpServletRequest request) {
-        if (request.getCookies() != null) {
-            for (Cookie cookie : request.getCookies()) {
-                if (REFRESH_TOKEN_COOKIE.equals(cookie.getName())) {
-                    return cookie.getValue();
-                }
-            }
-        }
-        throw new GeneralException(ErrorStatus.AUTH_REFRESH_TOKEN_NOT_FOUND);
-    }
 }
