+++ src/main/java/com/example/demo/security/filter/JwtAuthenticationFilter.java
@@ -1,12 +1,18 @@
 package com.example.demo.security.filter;
 
+import com.example.demo.common.exception.ErrorStatus;
+import com.example.demo.common.exception.GeneralException;
+import com.example.demo.common.util.CookieUtil;
+import com.example.demo.security.jwt.dto.JwtToken;
 import com.example.demo.security.jwt.service.TokenService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
+import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpHeaders;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
@@ -15,12 +21,15 @@ import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
 
+import static com.example.demo.common.consts.StaticVariable.REFRESH_TOKEN_COOKIE;
+
 @Slf4j
 @RequiredArgsConstructor
 @Component
 public class JwtAuthenticationFilter extends OncePerRequestFilter {
 
     private final TokenService tokenService;
+    private final CookieUtil cookieUtil;
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
